Fail build on ERROR commit status & add logging to Jenkins output to mirror Slack notifications. (#11343)
* Add logging to Jenkins output to mirror Slack notifiations.

Was hard to debug why a job was acting as a "no-op". Logging was only
getting sent to Slack channel. In our case the commit status was "error"
and the job had logic to not go any further, but we didn't know that
until we looked at the Slack channel for the log message.

* Mark build as failure if commit status isn't SUCCESS.

* Call Jenkins variable in a more readable way.